load locale

+++ schedule_api_splatoon3ink/main.py
@@ -2,6 +2,8 @@ import os
 import sys
 import time
 import datetime
+import urllib.request
+import json
 
 import gspread
 
@@ -56,16 +58,19 @@ def main():
                'User-Agent': f'{config.user_agent};{urllib_version}'}
 
     # GET locale data from splatoon3.ink
-    url = f'{config.source_url}/locale/ko-KR.json'
-    req = urllib.request.Request(url=url,
-                                 headers=headers,
-                                 method='GET')
-    if DEBUG:
-        print(f'[{time.time()-STIME}] Request {url}')
-    with urllib.request.urlopen(req) as f:
-        locale = json.load(f.read().decode('utf-8'))
-    with open('ko-KR.locale', 'w') as f:
-        json.dumps(f, locale)
+    if not os.path.exists('ko-KR.locale'):
+        url = f'{config.source_url}/locale/ko-KR.json'
+        req = urllib.request.Request(url=url,
+                                     headers=headers,
+                                     method='GET')
+        if DEBUG:
+            print(f'[{time.time()-STIME}] Request {url}')
+        with urllib.request.urlopen(req) as f:
+            locale = json.loads(f.read().decode('utf-8'))
+        with open('ko-KR.locale', 'w') as f:
+            json.dump(locale, f)
+    with open('ko-KR.locale' ,'r') as f:
+        locale = json.load(f)
